// src/types/librarycard.type.ts
export interface Ilibrary {
    id: number;
    name: string;
    image: string;
    muscleGroups: string[];
    equipment: string;
    difficulty: "Beginner" | "Intermediate" | "Advanced";
    duration: number;
    caloriesBurned: number;
    sets: number;
    reps: string;
    rating: number;
    description: string;
    instructions: string[];
};